refactor: extract WorkoutProgress component - address Sandy Metz feedback
- Extract workout progress visualization and exercise info into focused component
- Create WorkoutProgress component (189 lines) with clear props interface
- Reduce main component from 767 to 644 lines (-123 lines)
- Include: progress bar, muscle fatigue display, exercise info, exercise notes
- Total reduction: 867 → 644 lines (-223 lines, -26% reduction)
- Part of issue #27 Sandy Metz review feedback
