+++ components/workout-execution-experimental.tsx
@@ -16,6 +16,7 @@ import { ExerciseReplacementModal } from './exercise-replacement-modal'
 import { useRealTimeMuscleVolume } from '@/hooks/useRealTimeMuscleVolume'
 import { WorkoutHeader } from './workout-header'
 import { SetLoggingForm } from './set-logging-form'
+import { WorkoutProgress } from './workout-progress'
 
 interface WorkoutExercise {
   id: string
@@ -363,15 +364,6 @@ export function WorkoutExecutionExperimental() {
     return `${minutes}:${secs.toString().padStart(2, '0')}`
   }
 
-  const getDifficultyColor = (difficulty: string) => {
-    switch (difficulty.toLowerCase()) {
-      case 'beginner': return 'bg-green-500/20 text-green-400 border-green-500/30'
-      case 'intermediate': return 'bg-yellow-500/20 text-yellow-400 border-yellow-500/30'
-      case 'advanced': return 'bg-red-500/20 text-red-400 border-red-500/30'
-      default: return 'bg-gray-500/20 text-gray-400 border-gray-500/30'
-    }
-  }
-
   const getRPEColor = (rpe: number) => {
     if (rpe <= 3) return 'text-green-400'
     if (rpe <= 6) return 'text-yellow-400'
@@ -418,135 +410,20 @@ export function WorkoutExecutionExperimental() {
         isFinishDisabled={sets.length === 0}
       />
 
-      {/* Progress Bar */}
-      <div className="bg-[#2C2C2E] h-1">
-        <div 
-          className="bg-[#FF375F] h-1 transition-all duration-300"
-          style={{ width: `${((currentExerciseIndex + 1) / workoutQueue.length) * 100}%` }}
-        />
-      </div>
-
-      {/* Real-time Muscle Fatigue Visualization */}
-      <div className="p-4 pb-2">
-        <Card className="bg-[#1C1C1E] border-[#2C2C2E]">
-          <CardHeader className="pb-3">
-            <CardTitle className="text-sm text-[#A1A1A3] flex items-center">
-              <Zap className="h-4 w-4 mr-2" />
-              Real-time Muscle Fatigue
-            </CardTitle>
-          </CardHeader>
-          <CardContent className="pt-0">
-            <div className="grid grid-cols-3 gap-2" data-testid="muscle-fatigue-display">
-              {muscleVolumeData.summary.length > 0 ? (
-                muscleVolumeData.summary.slice(0, 6).map(({ muscle, intensity }) => (
-                  <div key={muscle} className="text-center">
-                    <div 
-                      className={`h-2 rounded-full mb-1 ${
-                        intensity === 'very_high' ? 'bg-red-500' :
-                        intensity === 'high' ? 'bg-orange-500' :
-                        intensity === 'medium' ? 'bg-yellow-500' :
-                        intensity === 'low' ? 'bg-green-500' :
-                        'bg-gray-500'
-                      }`}
-                      data-testid={`${muscle.toLowerCase()}-muscle-indicator`}
-                      data-intensity={intensity}
-                    />
-                    <span className="text-xs text-[#A1A1A3]">
-                      {muscle.replace('_', ' ')}
-                    </span>
-                  </div>
-                ))
-              ) : (
-                // Show placeholder muscle groups when no volume data yet
-                ['Chest', 'Shoulders', 'Back', 'Arms', 'Legs', 'Core'].map((muscle) => (
-                  <div key={muscle} className="text-center">
-                    <div 
-                      className="h-2 rounded-full mb-1 bg-gray-500"
-                      data-testid={`${muscle.toLowerCase()}-muscle-indicator`}
-                      data-intensity="none"
-                    />
-                    <span className="text-xs text-[#A1A1A3]">
-                      {muscle}
-                    </span>
-                  </div>
-                ))
-              )}
-            </div>
-          </CardContent>
-        </Card>
-      </div>
+      {/* Workout Progress */}
+      <WorkoutProgress
+        currentExerciseIndex={currentExerciseIndex}
+        workoutQueue={workoutQueue}
+        currentExercise={currentExercise}
+        muscleVolumeData={muscleVolumeData}
+        exerciseNotes={exerciseNotes}
+        showExerciseMenu={showExerciseMenu}
+        setShowExerciseMenu={setShowExerciseMenu}
+        setShowReplaceModal={setShowReplaceModal}
+        updateExerciseNotes={updateExerciseNotes}
+      />
 
       <div className="p-4">
-        {/* Current Exercise */}
-        <Card className="bg-[#1C1C1E] border-[#2C2C2E] mb-6">
-          <CardHeader>
-            <div className="flex items-center justify-between">
-              <div>
-                <CardTitle className="text-xl text-white">{currentExercise.name}</CardTitle>
-                <CardDescription className="flex items-center space-x-2 mt-1">
-                  <span className="text-[#A1A1A3]">{currentExercise.equipment}</span>
-                  <Badge 
-                    variant="outline" 
-                    className={`text-xs ${getDifficultyColor(currentExercise.difficulty)}`}
-                  >
-                    {currentExercise.difficulty}
-                  </Badge>
-                </CardDescription>
-              </div>
-              <div className="flex items-center space-x-2">
-                <div className="text-right">
-                  <p className="text-sm text-[#A1A1A3]">Exercise</p>
-                  <p className="text-lg font-bold text-white">
-                    {currentExerciseIndex + 1}/{workoutQueue.length}
-                  </p>
-                </div>
-                <Button
-                  variant="ghost"
-                  size="icon"
-                  onClick={() => setShowExerciseMenu(!showExerciseMenu)}
-                  className="text-[#A1A1A3] hover:text-white hover:bg-[#2C2C2E]"
-                  data-testid="exercise-menu"
-                >
-                  <MoreHorizontal className="h-5 w-5" />
-                </Button>
-              </div>
-            </div>
-            
-            {/* Exercise Menu */}
-            {showExerciseMenu && (
-              <div className="mt-4 p-3 bg-[#2C2C2E] rounded-lg space-y-2">
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  onClick={() => setShowReplaceModal(true)}
-                  className="w-full justify-start text-[#A1A1A3] hover:text-white"
-                >
-                  Replace Exercise
-                </Button>
-              </div>
-            )}
-          </CardHeader>
-        </Card>
-
-        {/* Exercise Notes */}
-        <Card className="bg-[#1C1C1E] border-[#2C2C2E] mb-6">
-          <CardHeader>
-            <CardTitle className="text-sm text-[#A1A1A3] flex items-center">
-              <NotebookPen className="h-4 w-4 mr-2" />
-              Exercise Notes
-            </CardTitle>
-          </CardHeader>
-          <CardContent>
-            <Textarea
-              placeholder="Form cues, technique notes..."
-              value={exerciseNotes[currentExercise?.id] || ''}
-              onChange={(e) => updateExerciseNotes(e.target.value)}
-              className="bg-[#2C2C2E] border-[#3C3C3E] text-white resize-none"
-              rows={2}
-              data-testid="exercise-notes"
-            />
-          </CardContent>
-        </Card>
 
         {/* Planned Sets Preview with Log All Sets */}
         {workoutPlan && getExercisePlannedSets(currentExercise.id).length > 0 && (
