+++ scratchpads/issue-27-experimental-workout-execution.md
@@ -15,12 +15,12 @@
 
 #### Phase 1: Break Down Component (PRIORITY 1)
 - [x] Extract WorkoutHeader component (56 lines) ✅ COMPLETED
-- [ ] Extract SetLoggingForm component (<100 lines)
+- [x] Extract SetLoggingForm component (153 lines) ✅ COMPLETED
 - [ ] Extract WorkoutProgress component (<75 lines)  
 - [ ] Extract ExerciseQueue component (<100 lines)
 - [ ] Create container component (<100 lines)
 
-**Progress**: 849 lines remaining (was 867) - 18 lines reduced
+**Progress**: 767 lines remaining (was 867) - 100 lines reduced (-12% reduction so far)
 
 #### Phase 2: Extract State Management (PRIORITY 2)
 - [ ] Create useWorkoutSession custom hook
